# openapi/eki-api-v0.1.yaml
openapi: 3.1.1
info:
  title: eKI API
  version: 0.10.0
  description: |
    KI-gestützte Sicherheitsprüfung für Drehbücher - Filmakademie Baden-Württemberg

    Die eKI ist eine modulare REST-API, die Drehbücher aus dem eProjekt-System entgegennimmt,
    KI-gestützt auf Sicherheitsrisiken analysiert und strukturierte Reports zurückliefert.

    ## Kernprinzipien
    - **Design-First:** Diese OpenAPI-Spezifikation ist die Single Source of Truth
    - **Processing-Only:** Keine persistente Speicherung von Drehbüchern/Reports (nur Audit-Metadaten)
    - **Write-Through:** Ergebnisse werden sofort an eProjekt zurückgespielt
    - **Minimalinvasiv:** Integration nur über REST, keine direkten DB-Zugriffe

    ## Meilenstein-Status
    - M01: Projektgerüst & OpenAPI ✓
    - M02: Parser Basis (FDX) ✓
    - M03: PDF & Streaming-Parsing ✓
    - M04: Risiko-Taxonomie & Scoring ✓
    - M05: Reports (JSON/PDF) & One-Shot-GET ✓
    - M06: LLM-Adapter (Mistral Cloud) & KB-Grundlage ✓
    - M07: Großdokument-Optimierung ✓
    - M08: Security/Privacy & Delete-on-Delivery ✓

  contact:
    name: Filmakademie Baden-Württemberg
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://eki-stage.filmakademie.de
    description: Staging environment
  - url: https://eki.filmakademie.de
    description: Production environment
...
paths:
...
  /v1/security/check:
    post:
      tags:
        - Security
      summary: Synchronous security check
      description: >-
        Perform a synchronous security check for scripts ≤1 MB or ≤50 scenes.
        Accepts JSON (base64-encoded script_content) or multipart/form-data file upload (.fdx, .pdf).
      operationId: securityCheckSync
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActorUserId'
        - $ref: '#/components/parameters/ActorProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityCheckRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultipartSecurityCheckRequest'
      responses:
        '200':
          description: Security check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncSecurityCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File exceeds maximum size (10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/security/check:async:
    post:
      tags:
        - Security
      summary: Asynchronous security check
      description: >-
        Start an asynchronous security check via Temporal workflow.
        Accepts JSON (base64-encoded script_content) or multipart/form-data file upload (.fdx, .pdf).
        Script content is AES-encrypted in Redis; only a reference key is passed to the workflow.
        Supports idempotency keys and configurable delivery mode (pull/push).
      operationId: securityCheckAsync
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ActorUserId'
        - $ref: '#/components/parameters/ActorProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsyncSecurityCheckRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultipartAsyncSecurityCheckRequest'
      responses:
        '202':
          description: Security check job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncSecurityCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File exceeds maximum size (10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          description: Workflow engine temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
      description: >-
        Opaque API key transmitted as `Authorization: Bearer eki_<token>`.
        Keys are SHA-256 hashed and stored in PostgreSQL with expiry dates.
  parameters:
    ActorUserId:
      name: X-Actor-User-Id
      in: header
      description: Optional actor user ID for audit logging (must match authenticated API-key user if provided)
      required: false
      schema:
        type: string
    ActorProjectId:
      name: X-Actor-Project-Id
      in: header
      description: Actor project ID for audit logging
      required: false
      schema:
        type: string
  schemas:
...
    ScriptFormat:
      type: string
      enum: [fdx, pdf]
      description: "Supported script formats: Final Draft XML (.fdx) or PDF (.pdf)"
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
      description: Status of an async security check job
    RiskLevel:
      type: string
      enum: [critical, high, medium, low, info]
      description: >-
        Risk level classification, calculated from likelihood × impact:
        critical (≥16), high (≥10), medium (≥5), low (≥2), info (<2)
...
    SecurityCheckRequest:
      type: object
      required:
        - script_content
        - script_format
        - project_id
      properties:
        script_content:
          type: string
          description: Base64-encoded script content (FDX as UTF-8 text, PDF as binary)
          minLength: 1
          maxLength: 10485760
        script_format:
          $ref: '#/components/schemas/ScriptFormat'
        project_id:
          type: string
          description: eProjekt project ID
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]+$'
        script_id:
          type: integer
          nullable: true
          description: >-
            eProjekt document ID of the script/treatment being analyzed.
            Passed back to ePro on push delivery so ePro can correlate the
            assessment with the correct document version. Use -1 if unknown.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: >-
            Optional HTTPS callback URL for push delivery results.
            Must be on allowlisted eProjekt domains (epro.filmakademie.de,
            staging.epro.filmakademie.de). Private/internal IPs are blocked.
        delivery:
          type: string
          enum: [pull, push]
          default: pull
          description: >-
            Delivery mode: "pull" (One-Shot GET via /reports/{id}) or
            "push" (POST to ePro callback URL)
        idempotency_key:
          type: string
          nullable: true
          maxLength: 255
          description: >-
            Optional idempotency key to prevent duplicate jobs.
            If a job with the same key already exists, the existing job is returned.
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for audit trail (max 50 keys, values ≤1000 chars)
    AsyncSecurityCheckRequest:
      allOf:
        - $ref: '#/components/schemas/SecurityCheckRequest'
        - type: object
          properties:
            priority:
              type: integer
              minimum: 1
              maximum: 10
              default: 5
              description: Job priority (1=highest, 10=lowest)
    MultipartSecurityCheckRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: Script file upload (.fdx or .pdf, max 10 MB). PDF files are validated by magic bytes.
        project_id:
          type: string
          description: eProjekt project ID
        script_format:
          $ref: '#/components/schemas/ScriptFormat'
        script_id:
          type: integer
          nullable: true
          description: eProjekt document ID
        delivery:
          type: string
          enum: [pull, push]
          default: pull
        idempotency_key:
          type: string
          nullable: true
          maxLength: 255
    MultipartAsyncSecurityCheckRequest:
      allOf:
        - $ref: '#/components/schemas/MultipartSecurityCheckRequest'
        - type: object
          properties:
            priority:
              type: integer
              minimum: 1
              maximum: 10
              default: 5

    # --- Response Schemas ---
...
    MeasureItem:
      type: object
      required:
        - code
        - title
      properties:
        code:
          type: string
          description: "Measure code from the catalog (e.g. RIG-SAFETY, PSY-BRIEFING, INTIMACY-COORD)"
        title:
          type: string
          description: Human-readable measure title
        responsible:
          type: string
          description: "Responsible role (e.g. Stunt Coordination, Weapons Master, Production)"
        due:
          type: string
          description: "Due date template (e.g. pre-production, shooting-3d, shooting-0d)"
    RiskFinding:
      type: object
      required:
        - id
        - risk_level
        - category
        - description
        - recommendation
        - confidence
      properties:
        id:
          type: string
          description: Unique finding ID
        scene_number:
          type: string
          nullable: true
          description: Scene number where risk was found
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        category:
          type: string
          description: "Risk category: PHYSICAL, ENVIRONMENTAL, PSYCHOLOGICAL"
        risk_class:
          type: string
          description: >-
            Specific risk class from the taxonomy (e.g. FIRE, HEIGHT, INTIMACY,
            STUNTS, WEAPONS, WATER, VIOLENCE, TRAUMA). 23 classes across 3 categories.
        rule_id:
          type: string
          description: "Taxonomy rule ID (e.g. SEC-P-008 for FIRE, SEC-Y-003 for TRAUMA)"
        likelihood:
          type: integer
          minimum: 0
          maximum: 5
          description: "Likelihood of risk occurring (1-5, 0=unscored)"
        impact:
          type: integer
          minimum: 0
          maximum: 5
          description: "Severity of impact if risk occurs (1-5, 0=unscored)"
        description:
          type: string
          description: Human-readable description of the risk
        recommendation:
          type: string
          description: Mitigation recommendation
        measures:
          type: array
          items:
            $ref: '#/components/schemas/MeasureItem'
          description: Applicable safety measures from the codified catalog (20 measures)
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score of the finding (1.0 = FDX/certain, lower for PDF/LLM)
        line_reference:
          type: string
          nullable: true
          description: Specific line or dialogue reference in the script
        evidence:
          type: string
          description: Textual evidence or rule justification supporting the risk assessment
        vulnerability:
          type: string
          description: "Vulnerability factors (e.g. children, animals, stunt doubles)"
        complexity:
          type: string
          description: "Complexity factors (e.g. extras, multi-camera, water/fire elements)"
        exposure_duration:
          type: string
          description: "Estimated risk exposure duration (e.g. kurz, mehrere Stunden, ganzer Drehtag)"
    RiskSummary:
      type: object
      description: Count of findings per risk level
      required: [critical, high, medium, low, info]
      properties:
        critical:
          type: integer
          minimum: 0
        high:
          type: integer
          minimum: 0
        medium:
          type: integer
          minimum: 0
        low:
          type: integer
          minimum: 0
        info:
          type: integer
          minimum: 0
      example:
        critical: 1
        high: 3
        medium: 5
        low: 2
        info: 1
    SecurityReport:
      type: object
      required:
        - report_id
        - project_id
        - script_format
        - created_at
        - risk_summary
        - total_findings
        - findings
        - processing_time_seconds
      properties:
        report_id:
          type: string
          format: uuid
        project_id:
          type: string
        script_format:
          $ref: '#/components/schemas/ScriptFormat'
        created_at:
          type: string
          format: date-time
        risk_summary:
          $ref: '#/components/schemas/RiskSummary'
        total_findings:
          type: integer
        findings:
          type: array
          items:
            $ref: '#/components/schemas/RiskFinding'
        processing_time_seconds:
          type: number
          format: float
        metadata:
          type: object
          additionalProperties: true
    SyncSecurityCheckResponse:
      type: object
      required:
        - report
        - message
      properties:
        report:
          $ref: '#/components/schemas/SecurityReport'
        message:
          type: string
          example: Security check completed successfully
    AsyncSecurityCheckResponse:
      type: object
      required:
        - job_id
        - status
        - message
        - status_url
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        message:
          type: string
          example: "Security check job started (delivery=pull)"
        status_url:
          type: string
          description: URL to check job status (e.g. /v1/security/jobs/{job_id})
        estimated_completion_seconds:
          type: integer
          nullable: true
...
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
          nullable: true
        message:
          type: string
        error_code:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # -----------------------------------------------------------------------
    # M06: Knowledge Base schemas
    # -----------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad request (missing fields, invalid file type, empty file)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized (missing, invalid, or expired API key)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error (invalid base64, bad project_id format, SSRF blocked, etc.)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded (IP-based 60/min or API-key-based 1000/h)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'